requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
praw>=7.7.1
wikipedia>=1.4.0
pandas>=2.0.0
requests>=2.31.0
